scripts/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
c2pa-python
ctranslate2
decord
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader
faiss-cpu
fastapi
fastapi-cli
fastapi-cloud-cli
fastar
faster-whisper
gguf
hachoir
hf-xet
huggingface_hub
kokoro
librosa
moviepy
numpy
onnxruntime-gpu
openai
opencv-python-headless
openwakeword
pillow
PyAudio
qwen-vl-utils
sounddevice
soundfile
sentence_transformers
torch==2.10.0+cu126
torchaudio==2.10.0+cu126
torchcodec==0.10.0
torchvision==0.25.0+cu126
tqdm
transformers
ultralytics
uvicorn
websockets